Replace use of babel-register+nodemon with babel-node+babel-watch.
Thus compilation is done from the outside, not inside the process.
