+++ Procfile.dev
@@ -1,2 +1,2 @@
-web: node_modules/.bin/nodemon --debug index.js
+web: node_modules/.bin/babel-watch --debug 5858 index.js
 memcached: memcached -p 11211
