+++ package.json
@@ -9,6 +9,7 @@
     "npm": "3.8.x"
   },
   "scripts": {
+    "start": "yarn run dev",
     "verbose-dev": "DEBUG=verbose,info,error nf start --procfile Procfile.dev",
     "dev": "DEBUG=info,error nf start --procfile Procfile.dev",
     "staging": "DEBUG=info,error nf start --procfile Procfile.dev --env .env.staging",
@@ -65,6 +66,7 @@
     "babel-jest": "^16.0.0",
     "babel-plugin-rewire": "1.0.0-beta-2",
     "babel-plugin-transform-flow-strip-types": "^6.18.0",
+    "babel-watch": "^2.0.5",
     "danger": "^0.6.10",
     "deep-equal": "^1.0.1",
     "dotenv": "^2.0.0",
@@ -75,7 +77,6 @@
     "flow-bin": "^0.37.0",
     "foreman": "^1.4.1",
     "jest": "^17.0.3",
-    "nodemon": "^1.8.1",
     "sinon": "^1.17.2"
   },
   "jest": {
